add show personal information from db on personal center page
