--- src/main/java/cmpt276/demo/models/UserProfile.java
@@ -1,78 +1,0 @@
-package cmpt276.demo.models;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name="user_profile")
-public class UserProfile {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int pid;
-
-    private String phoneNumber;
-    private String address;
-    private String postCode;
-    private String email;
-
-    public UserProfile() {
-    }
-
-    @OneToOne(mappedBy = "userProfile")
-    private User user;
-
-    public UserProfile(User user) {
-        this.user = user;
-    }
-
-    public UserProfile(String email) {
-        this.email = email;
-    }
-
-    public int getPid() {
-        return pid;
-    }
-
-    public void setPid(int pid) {
-        this.pid = pid;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getPostCode() {
-        return postCode;
-    }
-
-    public void setPostCode(String postCode) {
-        this.postCode = postCode;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-}
