+++ src/main/resources/static/css/adduser.css
@@ -1,0 +1,77 @@
+
+body {
+    font-family: Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+    background-color: #f4f4f4;
+}
+
+#top-bar {
+    background-color: #333;
+    color: #ffffff;
+    padding: 10px 0;
+    text-align: center;
+}
+
+#navbar {
+    width: 250px;
+    height: 100vh;
+    background-color: #555;
+    float: left;
+    list-style-type: none;
+    padding: 0;
+}
+
+#navbar li {
+    padding: 10px 20px;
+    border-bottom: 1px solid #666;
+}
+
+#navbar li a {
+    text-decoration: none;
+    color: #fff;
+    display: block;
+    transition: background-color 0.3s;
+}
+
+#navbar li a:hover {
+    background-color: #666;
+}
+
+#address-book {
+    border: 1px solid #ddd;
+    width: 60%;
+    margin: 0 auto;
+    padding: 20px;
+    border-radius: 10px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+#search-bar {
+    margin-bottom: 20px;
+}
+
+table {
+    width: 100%;
+    border-collapse: collapse;
+}
+
+table, th, td {
+    border: 1px solid #ddd;
+}
+
+th, td {
+    padding: 8px 15px;
+    text-align: left;
+}
+
+th {
+    background-color: #f4f4f4;
+}
+
+tr:hover {
+    background-color: #f6f6f6;
+}
+.input-box input{
+    width: 200px;height: 30px
+}
