+++ src/main/resources/static/css/dashboard.css
