+++ src/main/resources/templates/users/Settings.html
@@ -25,7 +25,7 @@
         <li><a href="#">Training</a></li>
         <li><a href="/users/personalCenter">Personal Center</a></li>
         <li><a href="/users/settings">Settings</a></li>
-        <li><a href="#">ffff</a></li>
+        <li><a href="/logout">logout</a></li>
     </ul>
