--- src/main/resources/templates/users/addedUser.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Register success</title>
-</head>
-<body>
-    Register success!!
-</body>
-</html>
