+++ src/main/resources/templates/users/dashboard.html
@@ -10,7 +10,7 @@
 
 <body>
     <div id="top-bar">
-        <h1> Enterprise Information Management System</h1> 
+        <h1> Enterprise Information Management System</h1>
     </div>
 
     <!-- menu-->
@@ -25,7 +25,7 @@
         <li><a href="#">Training</a></li>
         <li><a href="/users/personalCenter">Personal Center</a></li>
         <li><a href="/users/settings">Settings</a></li>
-        <li><a href="#">ffff</a></li>
+        <li><a href="/logout">logout</a></li>
     </ul>
 
     <div id="content">
