+++ src/main/resources/templates/users/personalCenter.html
@@ -27,7 +27,7 @@
         <li><a href="#">Training</a></li>
         <li><a href="/users/personalCenter">Personal Center</a></li>
         <li><a href="/users/settings">Settings</a></li>
-        <li><a href="#">ffff</a></li>
+        <li><a href="/logout">logout</a></li>
     </ul>
 
 
@@ -38,30 +38,27 @@
                 <img src="path_to_employee_image.jpg" alt="Employee Name">
             </div>
             <div>
-                <div class="profile-name">AAA</div>
-                <div class="profile-position">Software Engineer</div>
+                <div class="profile-name"   th:text="${user.username}">AAA</div>
+                <div class="profile-position" >Software Engineer</div>
             </div>
         </div>
         
         <!-- Personal Information -->
         <div class="profile-section">
             <span class="profile-label">Email:</span>
-            <span class="profile-value">AAA@gmail.com</span>
+            <span class="profile-value" th:text="${user.email}">AAA@gmail.com</span>
         </div>
         
         <div class="profile-section">
             <span class="profile-label">Phone Number:</span>
-            <span class="profile-value">123-456-7890</span>
+            <span class="profile-value" th:text="${user.phoneNumber}">123-456-7890</span>
         </div>
         
-        <div class="profile-section">
-            <span class="profile-label">Address:</span>
-            <span class="profile-value">123 Street, City, Country</span>
-        </div>
+
         
         <div class="profile-section">
             <span class="profile-label">Department:</span>
-            <span class="profile-value">Engineering</span>
+            <span class="profile-value" th:text="${user.department}">Engineering</span>
         </div>
         
         <!-- Add more sections as needed -->
